+++ kamehouse-commons-core/src/main/java/com/nicobrest/kamehouse/commons/model/systemcommand/JvncSenderSystemCommand.java
@@ -1,5 +1,6 @@
 package com.nicobrest.kamehouse.commons.model.systemcommand;
 
+import be.jedi.jvncsender.VncSender;
 import com.nicobrest.kamehouse.commons.exception.KameHouseInvalidDataException;
 import com.nicobrest.kamehouse.commons.utils.DockerUtils;
 import com.nicobrest.kamehouse.commons.utils.EncryptionUtils;
@@ -7,13 +8,17 @@ import com.nicobrest.kamehouse.commons.utils.FileUtils;
 import com.nicobrest.kamehouse.commons.utils.JsonUtils;
 import com.nicobrest.kamehouse.commons.utils.PropertiesUtils;
 import com.nicobrest.kamehouse.commons.utils.StringUtils;
+import java.util.List;
+import java.util.concurrent.TimeUnit;
 
 /**
  * JvncSender kamehouse-cmd command.
  *
  * @author nbrest
  */
-public class JvncSenderSystemCommand extends KameHouseCmdSystemCommand {
+public class JvncSenderSystemCommand extends SystemCommand {
+
+  private static final int VNC_PORT = 5900;
 
   private String text;
 
@@ -23,16 +28,43 @@ public class JvncSenderSystemCommand extends KameHouseCmdSystemCommand {
   public JvncSenderSystemCommand(String text) {
     logCommand = false;
     this.text = text;
-    setWindowsCmdStartMinimized(true);
-    setKameHouseCmdCommands();
+    setOutputCommand();
+  }
+
+  /**
+   * Setup jvncsender system command.
+   */
+  public JvncSenderSystemCommand(String text, int sleepTime) {
+    logCommand = false;
+    setSleepTime(sleepTime);
+    this.text = text;
+    setOutputCommand();
   }
 
   @Override
-  protected String getKameHouseCmdArguments() {
+  public SystemCommand.Output execute() {
     String host = DockerUtils.getHostname();
     String password = getVncServerPassword();
-    return "-o jvncsender -host \"" + host + "\" -port 5900 -password \"" + password + "\" -text \""
-        + text + "\"";
+    logger.info("Sending text to vnc server {}:{}", host, VNC_PORT);
+    try {
+      VncSender vncSender = new VncSender(host, VNC_PORT, password);
+      vncSender.sendText(text);
+      output.setStandardOutput(List.of("VNC command executed successfully"));
+    } catch (Exception e) {
+      logger.error("Error sending text to vnc server", e);
+      output.setStandardError(List.of("Error executing VNC command", e.getMessage()));
+    }
+    try {
+      int sleepTime = getSleepTime();
+      if (sleepTime > 0) {
+        logger.debug("Sleeping for {} seconds", sleepTime);
+        TimeUnit.SECONDS.sleep(sleepTime);
+      }
+    } catch (InterruptedException e) {
+      logger.warn("Interrupted exception", e);
+      Thread.currentThread().interrupt();
+    }
+    return output;
   }
 
   /**