+++ changelog.md
@@ -15,6 +15,7 @@
 - Moved screen controller buttons to a separate page from server management
 - Added ESC, ENTER and ALT+TAB, WIN+TAB and right arrow key press buttons in the screen controller page
 - Updated jvncsender to version 1.02
+- Moved jvncsender commands to execute direcly from webapps rather than going through kamehouse-cmd
 - Fixed broken execute on docker host commands
 - Fixed bugs